+++ bin/blocks/carousel.sh
@@ -1,0 +1,62 @@
+#!/bin/bash
+# @Author: Roni Laukkarinen
+# @Date:   2022-02-10 10:44:02
+# @Last Modified by:   Roni Laukkarinen
+# @Last Modified time: 2022-02-10 14:05:12
+
+
+# // New files/Dependencies (this file will install them)::
+# // ├── sass/gutenberg/blocks/_carousel.scss (automatic from get-block.sh)
+# // ├── node_modules/slick-carousel (automatic from theme npm)
+# // ├── sass/features/_slick.scss
+# // ├── js/src/gutenberg-editor.js
+# // ├── svg/slider-left-arrow.svg
+# // ├── svg/slider-right-arrow.svg
+# // └── svg/block-icons/carousel.svg (automatic from get-block.sh)
+
+# // Changes to files/folders:
+# // ├── js/src/front-end.js
+# // ├── sass/gutenberg/_blocks.scss
+# // ├── sass/global.scss
+# // ├── acf-json/
+# // └── functions.php
+
+# Block specific variables
+export BLOCK_ACF_JSON_FILE="group_61b9a58583b15.json"
+export BLOCK_ACF_JSON_PATH="${AIRBLOCKS_THEME_PATH}/acf-json/${BLOCK_ACF_JSON_FILE}"
+
+# SCSS features
+cp -nv ${AIRBLOCKS_THEME_PATH}/sass/features/_slick.scss ${PROJECT_THEME_PATH}/sass/features/
+
+# Import feature styles to global.scss
+sed -e "/\/\/ Global features/a\\
+@import 'features/slick';" < ${PROJECT_THEME_PATH}/sass/global.scss > ${PROJECT_THEME_PATH}/sass/global_with_changes.scss
+rm ${PROJECT_THEME_PATH}/sass/global.scss
+mv ${PROJECT_THEME_PATH}/sass/global_with_changes.scss ${PROJECT_THEME_PATH}/sass/global.scss
+
+# JS direct replaces
+LC_ALL=C sed -i '' -e "s;\/\/ import slick from \'slick-carousel\'\;;import slick from \'slick-carousel\'\;;" ${PROJECT_THEME_PATH}/js/src/front-end.js
+
+# Import js modules right after the last default js module in the front-end.js file
+sed -e "/\import \'what-input\'\;/a\\
+import './modules/accordion';" < ${PROJECT_THEME_PATH}/js/src/front-end.js > ${PROJECT_THEME_PATH}/js/src/front-end-with-changes.js
+rm ${PROJECT_THEME_PATH}/js/src/front-end.js
+mv ${PROJECT_THEME_PATH}/js/src/front-end-with-changes.js ${PROJECT_THEME_PATH}/js/src/front-end.js
+
+# Import js modules right after the last default js module in the gutenberg-editor.js file
+sed -e "/\import\/no-unresolved \*\//a\\
+import slick from 'slick-carousel';" < ${PROJECT_THEME_PATH}/js/src/gutenberg-editor.js > ${PROJECT_THEME_PATH}/js/src/gutenberg-editor-with-changes.js
+rm ${PROJECT_THEME_PATH}/js/src/gutenberg-editor.js
+mv ${PROJECT_THEME_PATH}/js/src/gutenberg-editor-with-changes.js ${PROJECT_THEME_PATH}/js/src/gutenberg-editor.js
+
+# Other SVG icons needed by this block
+cp -nv ${AIRBLOCKS_THEME_PATH}/svg/slider-left-arrow ${PROJECT_THEME_PATH}/svg/
+cp -nv ${AIRBLOCKS_THEME_PATH}/svg/slider-right-arrow.svg ${PROJECT_THEME_PATH}/svg/
+
+# Register ACF block in functions.php
+# Please note: The title of the block will be translated in localization.sh if en is selected
+sed -e "/\'acf_blocks\' \=\> \[/a\\
+      [|\
+       'name' => 'carousel',|\
+       'title' => 'Kuvakaruselli',|\
+      ],\\" < ${PROJECT_THEME_PATH}/functions.php | tr '|' '\n' > ${PROJECT_THEME_PATH}/tmpfile